+++ PokeAlarm/Alarms/Discord/__init__.py
@@ -1,7 +1,7 @@
 try:
     import requests
 except ImportError:
-    from ..Utils import pip_install
+    from PokeAlarm.Utils import pip_install
 
     pip_install('requests')
