+++ PokeAlarm/Alarms/Slack/__init__.py
@@ -1,7 +1,7 @@
 try:
     import slacker
 except ImportError:
-    from ..Utils import pip_install
+    from PokeAlarm.Utils import pip_install
 
     pip_install('slacker', '0.9.24')
