+++ PokeAlarm/Alarms/Telegram/Stickers.py
