+++ PokeAlarm/Alarms/Twilio/__init__.py
@@ -1,7 +1,7 @@
 try:
     from twilio.rest import TwilioRestClient
 except ImportError:
-    from ..Utils import pip_install
+    from PokeAlarm.Utils import pip_install
 
     pip_install('twilio', '5.4.0')
