+++ PokeAlarm/Alarms/Twitter/__init__.py
@@ -1,7 +1,7 @@
 try:
     import twitter
 except ImportError:
-    from ..Utils import pip_install
+    from PokeAlarm.Utils import pip_install
 
     pip_install('twitter', '1.17.1')
