+++ PokeAlarm/Alarms/__init__.py
@@ -1,0 +1,1 @@
+from Alarm import Alarm
